refactor: separate logic into files
